+++ ChangeLog.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Fixed
+
+- fixed `ArangoOperations#upsert(T, UpsertStrategy)` (issue #92)
+  - Check `Persistable#isNew`
+- fixed `ArangoOperations#upsert(Iterable<T>, UpsertStrategy)` (issue #92)
+  - Check `Persistable#isNew`
+
 ## [1.2.0] - 2018-07-02
 
 ### Added